refactored form
